pyproject.toml
[project]
name = "onboard"
version = "0.1.0"
requires-python = ">=3.12, <=3.13"
description = "Onboard Flask application"
authors = [{ name = "Mike Glenn", email = "mglenn@ilude.com" }]
dependencies = [
    "apscheduler",
    "bs4",
    "docker",
    "feedparser",
    "flask",
    "flask[async]",
    "flask-assets",
    "flask-caching",
    "flask-minify",
    "flask-minify[go]",
    "gunicorn",
    "hypercorn==0.15.0",
    "langchain",
    "langchain-community",
    "langchain_core",
    "pandas",
    "python-dateutil",
    "python-dotenv",
    "pyyaml",
    "requests",
    "unidecode",
    "cssmin>=0.2.0",
    "sentence-transformers>=5.1.2",
    "chromadb>=1.2.1",
    "readability-lxml>=0.8.4.1",
]

[tool.deptry.per_rule_ignores]
DEP002 = ["gunicorn", "python-dotenv"]

[tool.black]
line-length = 88
skip-string-normalization = true
target-version = ['py312']

[dependency-groups]
dev = [
    "behave",
    "black>=23.12.1",
    "bottleneck>=1.5.0",
    "docutils>=0.22",
    "duckdb>=1.3.2",
    "graphviz>=0.21",
    "h5py>=3.14.0",
    "imutils>=0.5.4",
    "ipykernel>=6.30.1",
    "ipython>=9.4.0",
    "isort>=5.13.2",
    "jupyter>=1.1.1",
    "jupyterhub>=5.3.0",
    "keras>=3.11.2",
    "matplotlib>=3.10.5",
    "nltk>=3.9.1",
    "notebook>=7.4.5",
    "numpy>=2.1.3",
    "pandas-datareader>=0.10.0",
    "plotly>=6.3.0",
    "pyautogui>=0.9.54",
    "pylint>=3.3.8",
    "pyyaml>=6.0.2",
    "pytest>=7.4.3",
    "pytest-cov>=4.1.0",
    "pytest-flask>=1.3.0",
    "requests-cache>=1.2.1",
    "selenium",
    "webdriver-manager",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.43",
    "statsmodels>=0.14.5",
    "watermark>=2.5.0",
    "flake8>=7.0.0",
    "deptry>=0.23.1",
    "openai>=1.99.9",
]
